Refactor the refresh method from the debugger panel
