+++ src/devtools/client/debugger/src/actions/ui.js
@@ -11,12 +11,18 @@ import {
   getSource,
   getSourceContent,
   getFileSearchQuery,
+  selectedLocationHasScrolled,
+  getSelectedLocation,
+  getContext,
 } from "../selectors";
-import { selectSource } from "../actions/sources/select";
+import { selectSource, selectLocation } from "../actions/sources/select";
 import { getEditor, getLocationsInViewport } from "../utils/editor";
 import { searchContents } from "./file-search";
 import { copyToTheClipboard } from "../utils/clipboard";
 import { isFulfilled } from "../utils/async-value";
+
+import { getCodeMirror } from "devtools/client/debugger/src/utils/editor";
+import { resizeBreakpointGutter } from "../utils/ui";
 
 export function setPrimaryPaneTab(tabName) {
   return { type: "SET_PRIMARY_PANE_TAB", tabName };
@@ -154,5 +160,44 @@ export function copyToClipboard(source) {
     if (content && isFulfilled(content) && content.value.type === "text") {
       copyToTheClipboard(content.value.value);
     }
+  };
+}
+
+export function refreshCodeMirror() {
+  return ({ dispatch, getState }) => {
+    // CodeMirror does not update properly when it is hidden. This method has
+    // a few workarounds to get the editor to behave as expected when switching
+    // to the debugger from another panel and the selected location has changed.
+    const codeMirror = getCodeMirror();
+
+    if (!codeMirror) {
+      return;
+    }
+
+    // Update CodeMirror by dispatching a resize event to the window. CodeMirror
+    // also has a refresh() method but it did not work as expected when testing.
+    window.dispatchEvent(new Event("resize"));
+
+    // After CodeMirror refreshes, scroll it to the selected location, unless
+    // the user explicitly scrolled the editor since the location was selected.
+    // In this case the editor will already be in the correct state, and we
+    // don't want to undo the scrolling which the user did.
+    const handler = () => {
+      codeMirror.off("refresh", handler);
+      setTimeout(() => {
+        const hasScrolled = selectedLocationHasScrolled(getState());
+        if (!hasScrolled) {
+          const location = getSelectedLocation(getState());
+          const cx = getContext(getState());
+
+          if (location) {
+            dispatch(selectLocation(cx, location));
+          }
+        }
+        resizeBreakpointGutter(codeMirror);
+        codeMirror.refresh();
+      }, 0);
+    };
+    codeMirror.on("refresh", handler);
   };
 }
